- Fixed some warnings. - Added custom autoexp.dat file.
git-svn-id: https://fuji.googlecode.com/svn/trunk@756 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Source/FileSystem/MFFileSystemCachedFile.cpp
@@ -150,7 +150,7 @@ int MFFileCachedFile_Read(MFFile* pFile, void *pBuffer, int64 bytes)
 
 			MFFile_Seek(pCacheData->pBaseFile, (int)pBucket->fileOffset, MFSeek_Begin);
 			uint32 read = MFFile_Read(pCacheData->pBaseFile, pBucket->pData, pBucket->size);
-			MFDebug_Assert(read == pBucket->size, "Error reading base file...");
+			MFDebug_Assert(read == (uint32)pBucket->size, "Error reading base file...");
 		}
 
 		pBucket->lastTouched = (uint32)MFSystem_ReadRTC();
